Refactoring of Parser class
The refactorinf of Parser class consist on decomposion of this
class into 3 new types: Content, FileContent and NonUnicode.

Content: is a interface the represent a concept: a content;

FileContent: implements the content concept and apply it to a file;

NonUnicode: implements the content concept and performs like a
filter, filtering all non unicode characters from a content.